requirements.txt
# Multi File RAG Dependencies


# Web Framework
flask>=2.3.0
flask-cors>=4.0.0
werkzeug>=2.3.0

# Data Processing
numpy>=1.24.0

# LangChain & Document Loaders (Must Have)
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1

# Document Processing
pypdf>=3.0.0
python-docx>=1.0.0
docx2txt>=0.8

# Multimodal PDF Processing
pymupdf>=1.23.0

# Embeddings (Must Have)
sentence-transformers>=2.2.0

# Optional: Audio Transcription
# openai-whisper>=20231117

# Utilities
python-magic>=0.4.27

accelerate==1.12.0
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
Deprecated==1.3.1
emoji==2.15.0
executing==2.2.1
filelock==3.24.0
filetype==1.2.0
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.2.0
google-api-core==2.29.0
google-auth==2.48.0
google-cloud-vision==3.12.1
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
hf-xet==1.2.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
ipykernel==7.2.0
ipyplot==1.1.2
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
langdetect==1.0.9
llvmlite==0.46.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.2
numba==0.63.1
numpy==2.2.0
olefile==0.47
onnx==1.20.1
onnxruntime==1.24.1
opencv-python==4.13.0.92
packaging==26.0
pandas==3.0.0
parso==0.8.6
pdf2image==1.17.0
pdfminer.six==20260107
pi_heif==1.2.0
pikepdf==10.3.0
pillow==12.1.1
platformdirs==4.9.1
prompt_toolkit==3.0.52
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
pure_eval==0.2.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pypdf==6.7.0
pypdfium2==5.4.0
python-dateutil==2.9.0.post0
python-iso639==2026.1.31
python-magic==0.4.27
python-multipart==0.0.22
python-oxmsg==0.0.2
PyYAML==6.0.3
pyzmq==27.1.0
RapidFuzz==3.14.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.2
rsa==4.9.1
safetensors==0.7.0
scipy==1.17.0
setuptools==82.0.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
soupsieve==2.8.3
stack-data==0.6.3
sympy==1.14.0
timm==1.0.24
tokenizers==0.22.2
torch==2.10.0
torchvision==0.25.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
transformers==5.1.0
typer==0.23.1
typer-slim==0.23.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
unstructured==0.20.2
unstructured-client==0.42.10
unstructured.pytesseract==0.3.15
unstructured_inference==1.2.0
urllib3==2.6.3
wcwidth==0.6.0
webencodings==0.5.1
wrapt==2.1.1

